Fix various warnings

+++ schittkowski/problem_40.cc
@@ -182,7 +182,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem40)
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
-  solver_t::problem_t::scales_t scales (g->outputSize (), 1.);
+  solver_t::problem_t::scales_t scales
+    (static_cast<std::size_t> (g->outputSize ()), 1.);
 
   problem.addConstraint (g, intervals, scales);
