+++ schittkowski/problem_50.cc
@@ -188,7 +188,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem50)
   intervals.push_back (G<functionType_t>::makeInterval (6., 6.));
   intervals.push_back (G<functionType_t>::makeInterval (6., 6.));
   intervals.push_back (G<functionType_t>::makeInterval (6., 6.));
-  solver_t::problem_t::scales_t scales (g->outputSize (), 1.);
+  solver_t::problem_t::scales_t scales
+    (static_cast<std::size_t> (g->outputSize ()), 1.);
 
   problem.addConstraint (g, intervals, scales);
