+++ schittkowski/problem_53.cc
@@ -179,7 +179,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem53)
   solver_t::problem_t problem (f);
 
   for (F<functionType_t>::size_type i = 0; i < f.inputSize (); ++i)
-    problem.argumentBounds ()[i] = F<functionType_t>::makeInterval (-10., 10.);
+    problem.argumentBounds ()[static_cast<std::size_t> (i)] =
+      F<functionType_t>::makeInterval (-10., 10.);
 
   boost::shared_ptr<G<functionType_t> > g =
     boost::make_shared<G<functionType_t> > ();
@@ -188,7 +189,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem53)
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
   intervals.push_back (G<functionType_t>::makeInterval (0., 0.));
-  solver_t::problem_t::scales_t scales (g->outputSize (), 1.);
+  solver_t::problem_t::scales_t scales
+    (static_cast<std::size_t> (g->outputSize ()), 1.);
 
   problem.addConstraint (g, intervals, scales);
