+++ schittkowski/problem_72.cc
@@ -172,7 +172,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem72)
   solver_t::problem_t problem (f);
 
   for (F<functionType_t>::size_type i = 0; i < f.inputSize (); ++i)
-    problem.argumentBounds ()[i] = F<functionType_t>::makeInterval
+    problem.argumentBounds ()[static_cast<std::size_t> (i)] =
+      F<functionType_t>::makeInterval
       (1e-3, (4. - static_cast<double> (i)) * 1e5);
 
   boost::shared_ptr<G<functionType_t> > g =
@@ -181,7 +182,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem72)
   solver_t::problem_t::intervals_t intervals;
   for (F<functionType_t>::size_type i = 0; i < g->outputSize (); ++i)
     intervals.push_back (G<functionType_t>::makeLowerInterval (0.));
-  solver_t::problem_t::scales_t scales (g->outputSize (), 1.);
+  solver_t::problem_t::scales_t scales
+    (static_cast<std::size_t> (g->outputSize ()), 1.);
 
   problem.addConstraint (g, intervals, scales);
